+++ src/lib/types/data/backend.ts
@@ -1,8 +1,8 @@
 export interface IBackendTranslation{
-     key_name: string,
-     base_string: string,
-     translation_string: string,
-     line_number: number
+    key_name: string,
+    base_string: string,
+    translation_string: string,
+    line_number: number
 }
 
 export interface CreateTranslationResult {